+++ http/src/main/scala/com/twitter/finatra/http/exceptions/MaxForwardsExceededException.scala
@@ -1,6 +1,12 @@
 package com.twitter.finatra.http.exceptions
 
-class MaxForwardsExceededException(maxDepth: Int) extends Exception {
+/**
+ * Denotes that Controller forwarding exceeded the maximum depth.
+ * @param maxDepth the maximum number of forwards allowed within the server.
+ * @see [[com.twitter.finatra.http.request.HttpForward]]
+ * @see [[com.twitter.finatra.http.routing.HttpRouter#withMaxRequestForwardingDepth]]
+ */
+class MaxForwardsExceededException private[finatra](maxDepth: Int) extends Exception {
   override def getMessage: String =
     s"Exceeded maximum depth of $maxDepth for nested requests"
 }